pyproject.toml
[tool.black]
line-length = 120
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
  | \.eggs
  | \.git
  | \.tox
  | \.venv
  | _build
  | build
  | changelog
  | dist
)/
'''

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = true
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = false
fail-under = 75
exclude = [
    "docs",
    "tests",
]
ignore-regex = ["Arguments", "Meta"]
whitelist-regex = []
verbose = 0
quiet = false
color = true

[tool.isort]
multi_line_output = 3
line_length = 120
known_first_party = [ "tests", "ded.utils" ]
include_trailing_comma = true
order_by_type = false
skip = [ ".docker", "app" ]

[tool.mypy]
plugins = "strawberry.ext.mypy_plugin"

[tool.poetry]
name = "ded"
version = "0.1.0"
description = ""
authors = ["Geoff Johnson <geoff.jay@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = {extras = ["debug-server"], version = "^0.93.3"}
supabase = "^0.0.3"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^20.8b1"
coverage = "^5.5"
interrogate = "^1.3.2"
isort = "^5.8.0"
pre-commit = "^2.12.1"
mypy = "^0.910"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
